+++ imod/mf6/rch.py
@@ -34,6 +34,7 @@ from imod.typing.grid import (
     enforce_dim_order,
     is_planar_grid,
 )
+from imod.util.dims import drop_layer_dim_cap_data
 
 
 class Recharge(BoundaryCondition, IRegridPackage):
@@ -271,7 +272,7 @@ class Recharge(BoundaryCondition, IRegridPackage):
         used to couple MODFLOW6 to MetaSWAP models. Active cells will have a
         recharge rate of 0.0.
         """
-        cap_data = cast(GridDataDict, imod5_data["cap"])
+        cap_data = cast(GridDataDict, drop_layer_dim_cap_data(imod5_data)["cap"])
 
         msw_area = get_cell_area_from_imod5_data(cap_data)
         msw_active = is_msw_active_cell(target_dis, cap_data, msw_area)