+++ imod/tests/fixtures/imod5_cap_data.py
@@ -21,14 +21,21 @@ def zeros_grid():
 
 @pytest.fixture(scope="function")
 def cap_data_sprinkling_grid() -> Imod5DataDict:
-    type = zeros_grid()
-    type[:, 1] = 1
-    type[:, 2] = 2
-    layer = xr.ones_like(type)
+    boundary = zeros_grid() + 1
+    wetted_area = zeros_grid() + 0.5
+    urban_area = zeros_grid() + 0.25
+
+    artificial_rch_type = zeros_grid()
+    artificial_rch_type[:, 1] = 1
+    artificial_rch_type[:, 2] = 2
+    layer = xr.ones_like(artificial_rch_type)
     layer[:, 1] = 2
 
     cap_data = {
-        "artificial_recharge": type,
+        "boundary": boundary,
+        "wetted_area": wetted_area,
+        "urban_area": urban_area,
+        "artificial_recharge": artificial_rch_type,
         "artificial_recharge_layer": layer,
         "artificial_recharge_capacity": xr.DataArray(25.0),
     }
@@ -37,10 +44,38 @@ def cap_data_sprinkling_grid() -> Imod5DataDict:
 
 
 @pytest.fixture(scope="function")
+def cap_data_sprinkling_grid_with_layer() -> Imod5DataDict:
+    boundary = zeros_grid() + 1
+    wetted_area = zeros_grid() + 0.5
+    urban_area = zeros_grid() + 0.25
+
+    artificial_rch_type = zeros_grid()
+    artificial_rch_type[:, 1] = 1
+    artificial_rch_type[:, 2] = 2
+    layer = xr.ones_like(artificial_rch_type)
+    layer[:, 1] = 2
+
+    cap_data = {
+        "boundary": boundary.expand_dims("layer"),
+        "wetted_area": wetted_area.expand_dims("layer"),
+        "urban_area": urban_area.expand_dims("layer"),
+        "artificial_recharge": artificial_rch_type.expand_dims("layer"),
+        "artificial_recharge_layer": layer.expand_dims("layer"),
+        "artificial_recharge_capacity": xr.DataArray(25.0),
+    }
+
+    return {"cap": cap_data}
+
+
+@pytest.fixture(scope="function")
 def cap_data_sprinkling_points() -> Imod5DataDict:
-    type = zeros_grid()
-    type[:, 1] = 3000
-    type[:, 2] = 4000
+    boundary = zeros_grid() + 1
+    wetted_area = zeros_grid() + 0.5
+    urban_area = zeros_grid() + 0.25
+
+    artificial_rch_type = zeros_grid()
+    artificial_rch_type[:, 1] = 3000
+    artificial_rch_type[:, 2] = 4000
 
     data = {
         "id": [3000, 4000],
@@ -52,7 +87,10 @@ def cap_data_sprinkling_points() -> Imod5DataDict:
 
     layer = pd.DataFrame(data=data)
     cap_data = {
-        "artificial_recharge": type,
+        "boundary": boundary,
+        "wetted_area": wetted_area,
+        "urban_area": urban_area,
+        "artificial_recharge": artificial_rch_type,
         "artificial_recharge_layer": layer,
         "artificial_recharge_capacity": xr.DataArray(25.0),
     }